Add lstdelone, it frees the link’s-content using the function del given as a parameter
Takes as a parameter a link’s pointer address and frees the
memory of the link’s content using the function del given as
a parameter, then frees the link’s memory using free(3). The
memory of next must not be freed.
Finally, the pointer to the link-conter that was just freed must be
set to NULL